refactor: payments modal 수정
