+++ src/components/Payments/PaymentsPayInfo/PaymentsPayInfo.tsx
