+++ src/pages/Payments/Payments.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useRef, useState } from "react";
 import Header from "@/components/Common/Header/Header";
 import MobCard from "@/components/Common/MobCard";
-import PaymentsGoods from "@/components/Payments/PaymentsGoods";
-import PaymentsPayInfo from "@/components/Payments/PaymentsPayInfo";
+import PaymentsGoods from "@/components/Payments/PaymentsGoods/PaymentsGoods";
+import PaymentsPayInfo from "@/components/Payments/PaymentsPayInfo/PaymentsPayInfo";
 import BottomButton from "@/components/Common/BottomButton/BottomButton";
 
 import {
@@ -12,11 +12,15 @@ import {
 import { nanoid } from "nanoid";
 
 import postPaymentData from "@/apis/postPaymentData";
+import { ModalLayer } from "@/components/Common/Modal/Modal";
+import useModal from "@/hooks/useModal";
+import { PaymentsModal } from "@/components/Payments/PaymentsModal/PaymentsModal";
 
 const Payments = () => {
   const [selectedPoint, setSelectedPoint] = useState<string>("");
   const [amount, setAmount] = useState(0);
-
+  const [modal, setModal] = useState(<></>);
+  const { openModal, closeModal, isModalOpened } = useModal();
   const clientKey = "test_ck_D5GePWvyJnrK0W0k6q8gLzN97Eoq";
   const customerKey = nanoid();
 
@@ -37,6 +41,14 @@ const Payments = () => {
   };
 
   useEffect(() => {
+    setModal(
+      <PaymentsModal
+        closeModal={closeModal}
+        text="현재 충전페이지는 테스트 모드입니다."
+        buttonText="닫기"
+      />
+    );
+    openModal();
     (async () => {
       const paymentWidget = await loadPaymentWidget(clientKey, customerKey);
 
@@ -68,6 +80,7 @@ const Payments = () => {
   }, [amount]);
   return (
     <>
+      {isModalOpened && <ModalLayer onClick={closeModal}>{modal}</ModalLayer>}
       <MobCard>
         <Header content="충전하기" />
         <div className="flex justify-center py-[6rem] w-full">
