+++ src/pages/Shop/Shop.tsx
@@ -1,31 +1,17 @@
 import MyPoint from "@/components/MyPage/MyPage/MyPoint/MyPoint";
 import MobShopFrameSlice from "@/components/Shop/ShopFrameList/MobShopFrameSlice/MobShopFrameSlice";
 import "@/pages/Shop/swiperCustomStyle.css";
-import { Suspense } from "react";
+import { Suspense, useEffect, useState } from "react";
 import ShopFrameList from "@/components/Shop/ShopFrameList/ShopFrameList";
 import ShopTicketList from "@/components/Shop/ShopTicketList/ShopTicketList";
 import MainHeader from "@/components/Common/MainHeader/MainHeader";
 import LoadingBox from "@/components/Common/Loading/LoadingBox/LoadingBox";
-import useModal from "@/hooks/useModal";
-import { ModalLayer } from "@/components/Common/Modal/Modal";
-import { ShopTestModal } from "@/components/Shop/ShopModal/ShopTestModal/ShopTestModal";
 
 const Shop = () => {
-
   const [viewportWidth, setViewportWidth] = useState(window.innerWidth);
-  const [modal, setModal] = useState(<></>);
-  const { openModal, closeModal, isModalOpened } = useModal();
   const isMobile = viewportWidth <= 393;
 
   useEffect(() => {
-    setModal(
-      <ShopTestModal
-        closeModal={closeModal}
-        text="현재 충전페이지는 개발 중입니다."
-        buttonText="확인하기"
-      />
-    );
-    openModal();
     const handleResize = () => {
       setViewportWidth(window.innerWidth);
     };
@@ -35,10 +21,8 @@ const Shop = () => {
     };
   }, []);
 
-
   return (
     <>
-      {isModalOpened && <ModalLayer onClick={closeModal}>{modal}</ModalLayer>}
       <MainHeader headerText="포인트 상점" />
       <div className="px-[2.2rem] pt-[6.7rem] w-full h-full _sm:px-[1.5rem] _sm:pt-[6rem]">
         <Suspense fallback={<LoadingBox />}>
